app/backend/requirements.txt
-i https://pypi.org/simple
aioredis==1.3.0
asgiref==3.2.2
async-timeout==3.0.1
attrs==19.3.0
autobahn==19.10.1
automat==0.8.0
cffi==1.13.0
channels-redis==2.4.0
channels==2.3.0
constantly==15.1.0
cryptography==2.8
daphne==2.3.0
django-rest-knox==4.1.0
django==2.2.6
djangorestframework-camel-case==1.1.1
djangorestframework==3.10.3
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
msgpack==0.6.2
pycparser==2.19
pyhamcrest==1.9.0
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
twisted==19.7.0
txaio==18.8.1
zope.interface==4.6.0
